+++ src/main/java/org/embl/mobie/viewer/projectcreator/ProjectCreatorHelper.java
@@ -7,7 +7,7 @@ import org.embl.mobie.io.ome.zarr.loaders.N5OMEZarrImageLoader;
 import org.embl.mobie.io.ome.zarr.readers.N5OmeZarrReader;
 import org.embl.mobie.viewer.Dataset;
 import org.embl.mobie.viewer.view.View;
-import org.embl.mobie.viewer.view.additionalviews.AdditionalViews;
+import org.embl.mobie.viewer.view.AdditionalViews;
 import org.embl.mobie.io.util.FileAndUrlUtils;
 import ij.IJ;
 import ij.ImagePlus;
