+++ src/test/java/tests/TestRemoteZebrafish.java
@@ -6,7 +6,7 @@ import org.embl.mobie.viewer.MoBIE;
 import org.embl.mobie.viewer.MoBIESettings;
 import org.embl.mobie.viewer.display.AnnotatedRegionDisplay;
 import org.embl.mobie.viewer.display.SegmentationSourceDisplay;
-import org.embl.mobie.viewer.view.additionalviews.AdditionalViewsLoader;
+import org.embl.mobie.viewer.view.AdditionalViewsLoader;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
